+++ src/charsed.c
@@ -437,7 +437,7 @@ static void Change(
 		break;
 
 	case YC_WEAPON:
-		b->gun = CLAMP_OPPOSITE(b->gun + d, 0, GUN_COUNT - 1);
+		b->gun = CLAMP_OPPOSITE((int)(b->gun + d), 0, GUN_COUNT - 1);
 		break;
 	}
 }
